// src/types/database.ts
export interface Database {
  public: {
    Tables: {
// ...
      goals: {
        Row: {
          id: string
          user_id: string
          title: string
          target_amount: number
          target_currency: 'USD' | 'LBP'
          current_amount: number
          deadline: string
          status: 'Active' | 'Completed' | 'Failed' | 'Overdue'
          created_at: string
          updated_at: string
        }
        Insert: {
          id?: string
          user_id: string
          title: string
          target_amount: number
          target_currency?: 'USD' | 'LBP'
          current_amount?: number
          deadline: string
          status?: 'Active' | 'Completed' | 'Failed' | 'Overdue'
          created_at?: string
          updated_at?: string
        }
        Update: {
          id?: string
          user_id?: string
          title?: string
          target_amount?: number
          target_currency?: 'USD' | 'LBP'
          current_amount?: number
          deadline?: string
          status?: 'Active' | 'Completed' | 'Failed' | 'Overdue'
          updated_at?: string
        }
        Relationships: never[]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}